\name{ApproxBayesMeta-package}
\alias{ApproxBayesMeta-package}
\alias{ApproxBayesMeta}
\docType{package}
\title{Please fill a text title description into the txt file TextTitle in the root of the package directory.}
\description{Please fill a text description into the txt file TextDescription in the root of the package directory.}
\details{
\tabular{ll}{
  Package: \tab ApproxBayesMeta\cr
  Type: \tab Package\cr
  Version: \tab 0.1-31-07-2014\cr
  License: \tab GPL (>=2)\cr }
Please fill a text overview into the txt file TextOverview in the root of the package directory.
}
\author{Paul Newcombe \email{paul.newcombe@mrc-bsu.cam.ac.uk}}
