% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CheckUniqueCovariates.R
\name{CheckUniqueCovariates}
\alias{CheckUniqueCovariates}
\title{Checks whether all covariates in a covariate matrix are unique}
\usage{
CheckUniqueCovariates(X)
}
\arguments{
\item{X}{a covariate matrix}
}
\value{
Boolean set to TRUE if all covariates are unique
}
\description{
This function checks whether all covariates in an X-matrix are unique.
}
\author{
Paul Newcombe
}

