% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimGenoMvn.R
\name{SimGenoMvn}
\alias{SimGenoMvn}
\title{Simulates a matrix of genotypes using a latent multivariate normal}
\usage{
SimGenoMvn(seed = 1, N, B = 1, P, MAF, rSq, normalise = F)
}
\arguments{
\item{seed}{Random number seed}

\item{N}{number of people}

\item{B}{number of snp-blocks}

\item{P}{number of SNPs in each block}

\item{MAF}{Minor Allele Frequency of the SNPs in each block (vector of same length
as the number of blocks)}

\item{rSq}{pair-wise correlation between the latent normal variables (vector of
same length as the number of blocks)}
}
\value{
A N by P matrix of 0/1/2 genotype codes
}
\description{
Simulates a matrix of genotypes utilising a latent multivariate normal
distribution. All `SNPs' have the same specified MAF, and the same
(also specified) pairwise correlations. Therehsolds consistent with
HWE for the specified MAF are applied on the resulting continuous
covariates, to create categorical `genotypes'.
}

