% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GenerateXtxPlugin.R
\name{GenerateXtxPlugin}
\alias{GenerateXtxPlugin}
\title{Generates the plug-in estimate of the unobserved X'X (or `B' in Yang and Visscher)}
\usage{
GenerateXtxPlugin(n, W, B = 1, block.indices = NULL, mafs = NULL)
}
\arguments{
\item{n}{Number of individuals the marginal statistics are calculated from}

\item{W}{Covariate matrix from external data}

\item{B}{The number of covariate blocks}

\item{block.indices}{Optional vector of block indices (if not provided,
identical block sizes are used generated from the function BlockIndices)}

\item{mafs}{Vector of Minor Allele Frequencies. If left as NULL, these are
estimated from W}
}
\value{
A list of block plug-in estimates for X'X
}
\description{
Generates the plug-in estimate of X'X required for the marginal
meta-analysis model. If a covariate matrix, W, is available from
an external population of identical size, W'W would simply be used
as the plug-in estimate for X'X. However, if W is from a population
of a different size, then it must be scaled accordingly, to create
`B' as described by Yang and Visscher.
}
\author{
Paul Newcombe
}

